// src/types/market-place.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export type Response = {
  results: Result[]
}

export type Result = {
  extensions: Extension[]
  pagingToken: null
  resultMetadata: ResultMetadatum[]
}

export type Extension = {
  publisher: Publisher
  extensionId: string
  extensionName: string
  displayName: string
  flags: ExtensionFlags
  lastUpdated: string
  publishedDate: string
  releaseDate: string
  shortDescription?: string
  versions: Version[]
  deploymentType: number
}

export type ExtensionFlags = {
  ValidatedPublic: 'validated, public'
  ValidatedPublicPreview: 'validated, public, preview'
  ValidatedPublicUnpublished: 'validated, public, unpublished'
  ValidatedPublicUnpublishedLocked: 'validated, public, unpublished, locked'
}

export type Publisher = {
  publisherId: string
  publisherName: string
  displayName: string
  flags: PublisherFlags
  domain: null | string
  isDomainVerified: boolean
}

export type PublisherFlags = {
  Verified: 'verified'
}

export type Version = {
  version: string
  flags: VersionFlags
  lastUpdated: string
  files: File[]
  properties: Property[]
  assetUri: string
  fallbackAssetUri: string
}

export type File = {
  assetType: AssetType
  source: string
}

export type AssetType = {
  MicrosoftVisualStudioCodeManifest: 'Microsoft.VisualStudio.Code.Manifest'
  MicrosoftVisualStudioServicesContentChangelog: 'Microsoft.VisualStudio.Services.Content.Changelog'
  MicrosoftVisualStudioServicesContentDetails: 'Microsoft.VisualStudio.Services.Content.Details'
  MicrosoftVisualStudioServicesContentLicense: 'Microsoft.VisualStudio.Services.Content.License'
  MicrosoftVisualStudioServicesIconsDefault: 'Microsoft.VisualStudio.Services.Icons.Default'
  MicrosoftVisualStudioServicesIconsSmall: 'Microsoft.VisualStudio.Services.Icons.Small'
  MicrosoftVisualStudioServicesVSIXPackage: 'Microsoft.VisualStudio.Services.VSIXPackage'
  MicrosoftVisualStudioServicesVsixManifest: 'Microsoft.VisualStudio.Services.VsixManifest'
}

export type VersionFlags = {
  Validated: 'validated'
}

export type Property = {
  key: Key
  value: string
}

export type Key = {
  MicrosoftVisualStudioCodeEngine: 'Microsoft.VisualStudio.Code.Engine'
  MicrosoftVisualStudioCodeExtensionDependencies: 'Microsoft.VisualStudio.Code.ExtensionDependencies'
  MicrosoftVisualStudioCodeExtensionKind: 'Microsoft.VisualStudio.Code.ExtensionKind'
  MicrosoftVisualStudioCodeExtensionPack: 'Microsoft.VisualStudio.Code.ExtensionPack'
  MicrosoftVisualStudioCodeLocalizedLanguages: 'Microsoft.VisualStudio.Code.LocalizedLanguages'
  MicrosoftVisualStudioServicesBrandingColor: 'Microsoft.VisualStudio.Services.Branding.Color'
  MicrosoftVisualStudioServicesBrandingTheme: 'Microsoft.VisualStudio.Services.Branding.Theme'
  MicrosoftVisualStudioServicesCustomerQnALink: 'Microsoft.VisualStudio.Services.CustomerQnALink'
  MicrosoftVisualStudioServicesEnableMarketplaceQnA: 'Microsoft.VisualStudio.Services.EnableMarketplaceQnA'
  MicrosoftVisualStudioServicesGitHubFlavoredMarkdown: 'Microsoft.VisualStudio.Services.GitHubFlavoredMarkdown'
  MicrosoftVisualStudioServicesLinksGetstarted: 'Microsoft.VisualStudio.Services.Links.Getstarted'
  MicrosoftVisualStudioServicesLinksGitHub: 'Microsoft.VisualStudio.Services.Links.GitHub'
  MicrosoftVisualStudioServicesLinksLearn: 'Microsoft.VisualStudio.Services.Links.Learn'
  MicrosoftVisualStudioServicesLinksRepository: 'Microsoft.VisualStudio.Services.Links.Repository'
  MicrosoftVisualStudioServicesLinksSource: 'Microsoft.VisualStudio.Services.Links.Source'
  MicrosoftVisualStudioServicesLinksSupport: 'Microsoft.VisualStudio.Services.Links.Support'
}

export type ResultMetadatum = {
  metadataType: string
  metadataItems: MetadataItem[]
}

export type MetadataItem = {
  name: string
  count: number
}